Move code into src/ directory
This makes testing using the standalone app work, because by default,
the standalone computer only has 1kb of disk space, so using the project
root as the computer path used all space which means saving files could
not be tested within the computer.

It's possible to modify the standalone app to increase the disk space,
but this is easier for now.